Refactor: move components to features/TodoList and features/shared
